# backend/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking Management System API - Full Documentation",
        "description": "Comprehensive API documentation for the Vehicle Parking Management System backend.",
        "version": "1.0.0",
        "contact": {
            "name": "Parking System Support",
            "email": "support@parkingsystem.example.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Development Server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/parking-slots": {
            "post": {
                "tags": [
                    "Parking Slots"
                ],
                "summary": "Create a parking slot (Admin)",
                "operationId": "createParkingSlot",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateParkingSlotRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parking slot created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "slot": {
                                            "$ref": "#/components/schemas/ParkingSlot"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "get": {
                "tags": [
                    "Parking Slots"
                ],
                "summary": "List parking slots (Admin/User)",
                "operationId": "listParkingSlots",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/PageQueryParam"
                    },
                    {
                        "$ref": "#/components/parameters/LimitQueryParam"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by slot number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "slot_number",
                                "size",
                                "vehicle_type",
                                "status",
                                "location",
                                "created_at"
                            ],
                            "default": "slot_number"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/OrderQueryParam"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/ParkingSlotStatusEnum"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Filter by size",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/VehicleSizeEnum"
                        }
                    },
                    {
                        "name": "vehicle_type",
                        "in": "query",
                        "description": "Filter by vehicle type",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/VehicleTypeEnum"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter by location",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SlotLocationEnum"
                        }
                    },
                    {
                        "name": "showAll",
                        "in": "query",
                        "description": "Admin flag to show all slots",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of parking slots.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParkingSlotListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/parking-slots/bulk": {
            "post": {
                "tags": [
                    "Parking Slots"
                ],
                "summary": "Bulk create parking slots (Admin)",
                "operationId": "bulkCreateParkingSlots",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkCreateParkingSlotsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Slots created via bulk.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/parking-slots/{id}": {
            "parameters": [
                {
                    "$ref": "#/components/parameters/IdPathParamUuid"
                }
            ],
            "get": {
                "tags": [
                    "Parking Slots"
                ],
                "summary": "Get parking slot by ID",
                "operationId": "getParkingSlotById",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot details.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParkingSlot"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "put": {
                "tags": [
                    "Parking Slots"
                ],
                "summary": "Update parking slot (Admin)",
                "operationId": "updateParkingSlot",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateParkingSlotRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Slot updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "slot": {
                                            "$ref": "#/components/schemas/ParkingSlot"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Parking Slots"
                ],
                "summary": "Delete parking slot (Admin)",
                "operationId": "deleteParkingSlot",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slot deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestError"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "ParkingSlotStatusEnum": {
                "type": "string",
                "enum": [
                    "AVAILABLE",
                    "UNAVAILABLE",
                    "MAINTENANCE"
                ]
            },
...
            "ParkingSlot": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "slot_number": {
                        "type": "string"
                    },
                    "size": {
                        "$ref": "#/components/schemas/VehicleSizeEnum"
                    },
                    "vehicle_type": {
                        "$ref": "#/components/schemas/VehicleTypeEnum"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ParkingSlotStatusEnum"
                    },
                    "location": {
                        "$ref": "#/components/schemas/SlotLocationEnum"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CreateParkingSlotRequest": {
                "type": "object",
                "required": [
                    "slot_number",
                    "size",
                    "vehicle_type"
                ],
                "properties": {
                    "slot_number": {
                        "type": "string"
                    },
                    "size": {
                        "$ref": "#/components/schemas/VehicleSizeEnum"
                    },
                    "vehicle_type": {
                        "$ref": "#/components/schemas/VehicleTypeEnum"
                    },
                    "location": {
                        "$ref": "#/components/schemas/SlotLocationEnum"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ParkingSlotStatusEnum",
                        "default": "AVAILABLE"
                    }
                }
            },
...
        },
        "parameters": {
...
            "PageQueryParam": {
                "name": "page",
                "in": "query",
                "description": "Page number",
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
            "LimitQueryParam": {
                "name": "limit",
                "in": "query",
                "description": "Items per page",
                "schema": {
                    "type": "integer",
                    "default": 10
                }
            },
            "OrderQueryParam": {
                "name": "order",
                "in": "query",
                "description": "Sort order",
                "schema": {
                    "type": "string",
                    "enum": [
                        "asc",
                        "desc"
                    ],
                    "default": "desc"
                }
            },
...
        },
        "responses": {
            "UnauthorizedError": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "ForbiddenError": {
                "description": "Forbidden",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
...
            "BadRequestError": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            }
        }
    }
}